Update motor code from uint to int

+++ motor.c
@@ -87,7 +87,7 @@ void initMotor(void) {
 	TPM0_C3SC |= (TPM_CnSC_ELSB(1) | TPM_CnSC_MSB(1));
 }
 
-void move_right_forward_wheel(uint16_t speed, uint8_t direction) {
+void move_right_forward_wheel(int speed, int direction) {
 	if (direction) {
 		TPM0_C0V = 0;
 		TPM0_C1V = speed;
@@ -97,7 +97,7 @@ void move_right_forward_wheel(uint16_t speed, uint8_t direction) {
 	}
 }
 
-void move_right_backward_wheel(uint16_t speed, uint8_t direction) {
+void move_right_backward_wheel(int speed, int direction) {
 	if (direction) {
 		TPM0_C4V = 0;
 		TPM0_C5V = speed;
@@ -107,7 +107,7 @@ void move_right_backward_wheel(uint16_t speed, uint8_t direction) {
 	}
 }
 
-void move_left_forward_wheel(uint16_t speed, uint8_t direction) {
+void move_left_forward_wheel(int speed, int direction) {
 	if (direction) {
 		TPM1_C0V = 0;
 		TPM1_C1V = speed;
@@ -117,7 +117,7 @@ void move_left_forward_wheel(uint16_t speed, uint8_t direction) {
 	}
 }
 
-void move_left_backward_wheel(uint16_t speed, uint8_t direction) {
+void move_left_backward_wheel(int speed, int direction) {
 	if (direction) {
 		TPM0_C2V = 0;
 		TPM0_C3V = speed;
@@ -127,35 +127,35 @@ void move_left_backward_wheel(uint16_t speed, uint8_t direction) {
 	}
 }
 
-void move_forward_or_backward(uint16_t speed, uint8_t direction) {
+void move_forward_or_backward(int speed, int direction) {
   move_right_forward_wheel(speed, direction);
 	move_right_backward_wheel(speed, direction);
   move_left_forward_wheel(speed, direction);
 	move_left_backward_wheel(speed, direction);
 }
 
-void turn_right_on_spot(uint16_t speed) {
+void turn_right_on_spot(int speed) {
 	move_right_forward_wheel(speed, 0);
 	move_right_backward_wheel(speed, 0);
 	move_left_forward_wheel(speed, 1);
 	move_left_backward_wheel(speed, 1);
 }
 
-void turn_left_on_spot(uint16_t speed) {
+void turn_left_on_spot(int speed) {
 	move_right_forward_wheel(speed, 1);
 	move_right_backward_wheel(speed, 1);
 	move_left_forward_wheel(speed, 0);
 	move_left_backward_wheel(speed, 0);
 }
 
-void turn_left_smooth(uint16_t base_speed, uint16_t turn_factor, uint8_t direction) {
+void turn_left_smooth(int base_speed, int turn_factor, int direction) {
   move_left_forward_wheel(base_speed - turn_factor, direction);
 	move_left_backward_wheel(base_speed - turn_factor, direction);
   move_right_forward_wheel(base_speed, direction);
 	move_right_backward_wheel(base_speed, direction);
 }
 
-void turn_right_smooth(uint16_t base_speed, uint16_t turn_factor, uint8_t direction) {
+void turn_right_smooth(int base_speed, int turn_factor, int direction) {
 	move_left_forward_wheel(base_speed, direction);
 	move_left_backward_wheel(base_speed, direction);
 	move_right_forward_wheel(base_speed - turn_factor, direction);
@@ -173,7 +173,6 @@ void stop_car() {
 void move(int base_speed, int turn_factor){
 	// Forward is Negative
 	// Left is Negative
-	
 	int fwd_direction;
 	if (base_speed < 0) {
 		base_speed = -base_speed;
